+++ packages/charts/src/chart_types/metric/renderer/dom/index.tsx
@@ -155,6 +155,7 @@ function Component({
             style.minValueFontSize,
             datum.valueIcon !== undefined,
             style.valueFontSize === 'fit',
+            style.fontFamily,
             textDimensions.metricSpacing.panelPadding,
           );
           acc.fittedValueFontSize = Math.min(acc.fittedValueFontSize, valueFontSize);