+++ packages/charts/src/chart_types/metric/renderer/dom/metric.tsx
@@ -204,6 +204,7 @@ export const Metric: React.FunctionComponent<{
     backgroundColor: hasTrend || hasProgressBar ? backgroundColor : datumWithInteractionColor.color,
     cursor: onElementClick ? 'pointer' : DEFAULT_CSS_CURSOR,
     borderColor: style.border,
+    fontFamily: style.fontFamily,
     '--echMetricPadding': `${metricSpacing.panelPadding}px`,
     '--echMetricTitleSubtitleGap': `${metricSpacing.titleSubtitleGap}px`,
     '--echMetricExtraPaddingTop': `${metricSpacing.extraPaddingTop}px`,